a paired down org init
I tried a few things over the past few days. Ultimately I discovered
that magit is necessary, vc just doesn't cut it (the cli is
better). This new setup is very simple, so it should work on all my
platforms with ease. I intend not to change it for a while.