Implements evil-mode
Using vim like behaviour in as many places as possible should reduce
the cost of context switching throughout my work day. Additionally,
thought evil mode does seem to require a lot of configuration, it is
nothing that fundamentally alters my workflow, so I could bear the
cost of losing it if need be. For this reason I do not want to use
Doom, which carries even more caonfiguration risk.