Reimplemented god-mode
I found myself once more regretting the need to use modifiers so
heavily. God-mode is not perfect, but it is an extremely light-weight
solution. Evil is too thick of a layer, Doom thicker still. Viper
requires much more configuration to avoid key chording, a lot of
custom bindings would need to be made. Hydra requires a similar amount
of configuration. Other modal packages are simply too much to learn.

God-mode allows me to use one set of keybindings, normal emacs ones.
It is just that the control and meta keys are optimized away. This
accomplishes the basic goal of avoiding keychording.