changes to work init for new organization system
This new organizational system eliminates the need for org-roam.
Temporal information is fully captured by timestamps, and I rely on
org-agenda log mode and inactive timestamp mode to serve as a sort of
journal. For an archive, I archive to a datetree in journal.org. I
also write journal entries to this datetree. I
will split this journal up as appropriate to performance
optimizations. For regular task management I will use staging.org, or
put tasks in project files in the projects directory as appropriate.
staging.org is the replacement for tasks.org, which is the file I
actually work in and manipulate regularly (not including project
files, which I may also work in as appropriate). stagin includes the
headers Reminders/Checklists, Tasks, Appointments, and Inbox. This
replaces the inbox.org file. It also provides a place for Appointments
to go.

Another change is that I no longer expect all WAITs to have deadlines.
This way my deadlines are reserved for meaningful due dates. Instead I
review my list of WAITs regularly and then decide if I want to follow
up or not.

The advantages of this new system are as follows:
1. No plugin dependencies, relies only on core org features.
2. Provides a central place to view information sorted
temporally (org-agenda).
3. Fewer files and plugins is more performant.
4. Less decisions to worry about. As long as there are timestamps and
tags as appropriate, I will find what I need via the agenda views.
5. Less crowded staging area. Instead of a grab bag of things to sort
through, there are clear categories.

My hope is that with this system there are no more necessary changes.
I lose the functionality of org-roam with and org-roam-dailies with
this change, but I do not think that I was benefiting from these at
all. Any changes from now on should only need to be small tweaks, like
functions to automate work, or the management of my lists of tags and
task states, but these are changes I can make without fear of losing
any data.